chore: nettoyage clippy + factorisation MetaField + accents help
Vague d'audit cleanup (B3, B4, C1, C2, C4, D).

B3+D - Hygiene clippy

cargo clippy --fix automatique :
- 5 redundant_closure dans archive/count.rs et archive/extractor.rs
  (`|x| func(x)` -> `func`)
- 7 deref_addrof dans native_player/registry.rs et archive/mod.rs
  (`&*guard` -> `&guard`)
- 1 unnecessary_cast dans native_player/platform.rs
- 1 redundant_at_rest_pattern dans archive/mod.rs
- 2 unused imports : `count_and_estimate_archive_entries_filtered`
  (archive/mod.rs) et `video::VideoMetadata` (lib.rs)
- items_after_test_module dans scanner/hash.rs

Manuel :
- manual_checked_division dans video/hash.rs:247 (`sum / count` ->
  `sum.checked_div(count).unwrap_or(0)`)
- too_many_arguments sur le helper `pf` des tests de filters.rs :
  ajout d'un `#[allow]` local (miroir de la signature de
  `passes_filters` qui a deja l'allow)
- field_reassign_with_default : 13 occurrences dans le module test
  de scanner/mod.rs - allow au niveau du module test (le pattern
  `let mut cfg = Default::default(); cfg.x = ...` est preferable
  ici pour eviter de dupliquer la signature `..Default::default()`
  dans chaque test).

Resultat : 0 warning clippy sur lib + lib test.

B4 - AGENTS.md

Reference a `count_archive_image_entries` corrigee (fonction supprimee
Phase 31, remplacee par `count_and_estimate_archive_image_entries`).

C1 - ComparatorBasicShell deplace

Utilise uniquement par ArchiveComparator. Deplacement en fonction
locale (non exportee) dans `ArchiveComparator.tsx`. Allege
`comparatorShared.tsx` qui ne garde que ce qui est vraiment partage.

C2 - Accents francais dans help/content.ts

Articles `audio-comparator-open`, `audio-comparator-keep` et le
paragraphe permissions de `notifications` corriges (apparait ->
apparaît, en-tete -> en-tête, cote a cote -> côté à côté, maitre ->
maître, metadonnees -> métadonnées, etc.).

C4 - MetaField generique

Composant `<MetaField label value valueClassName? valueStyle?>` extrait
dans `comparatorShared.tsx`. Remplace les blocs
`<div className="comparator-meta-row"><span ...label></span><span ...
value></span></div>` repetes dans :
- AudioComparator (1 ligne duree)
- ImageComparator (3 lignes dimensions/format/exif_date)
- VideoComparator (4 lignes dimensions/duree/codec/audio - le 4eme
  utilise valueClassName/valueStyle pour preserver le markup
  conditionnel du label dynamique et de la liste audio_tracks)

4 nouveaux tests TS dans comparatorShared.test.tsx : rendu label+value
basique, accepte ReactNode, applique valueClassName / valueStyle a la
cellule value uniquement.

Tests : 330 Rust (inchange) / 491 TS (+4) / tsc clean / 0 warning clippy.

+++ src-tauri/src/native_player/platform.rs
@@ -39,7 +39,7 @@ impl ParentHandle {
             RawWindowHandle::Win32(h) => ParentHandle::Win32(h.hwnd.get() as u64),
             // h.window est `c_ulong` : u64 sur Linux 64 bits, u32 sur Windows. On caste
             // explicitement pour eviter une erreur de compile cross-platform.
-            RawWindowHandle::Xlib(h) => ParentHandle::Xlib(h.window as u64),
+            RawWindowHandle::Xlib(h) => ParentHandle::Xlib(h.window),
             RawWindowHandle::Xcb(h) => ParentHandle::Xcb(h.window.get() as u64),
             _ => ParentHandle::Unsupported,
         }
